# docs/swagger.yaml
basePath: /api
definitions:
...
  model.LoginUserRequest:
    properties:
      id:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        type: string
    required:
    - id
    - password
    type: object
...
  model.RefreshTokenRequest:
    properties:
      refresh_token:
        maxLength: 100
        type: string
    required:
    - refresh_token
    type: object
  model.RegisterUserRequest:
    properties:
      id:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        type: string
    required:
    - id
    - name
    - password
    type: object
...
  model.UpdateUserRequest:
    properties:
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        type: string
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      name:
        type: string
      refresh_token:
        type: string
      token:
        type: string
      updated_at:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for managing users, contacts, and addresses with clean
    architecture principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Golang Clean Architecture API
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Logout the currently authenticated user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            properties:
              data:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account with ID, name, and password
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered user
          schema:
            properties:
              data:
                $ref: '#/definitions/model.UserResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      summary: Register a new user
      tags:
      - users
  /users/_current:
    get:
      consumes:
      - application/json
      description: Get the currently authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: Current user information
          schema:
            properties:
              data:
                $ref: '#/definitions/model.UserResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update the currently authenticated user's information (name and/or
        password)
      parameters:
      - description: User update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user
          schema:
            properties:
              data:
                $ref: '#/definitions/model.UserResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
  /users/_login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive access token
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in with token
          schema:
            properties:
              data:
                $ref: '#/definitions/model.UserResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      summary: User login
      tags:
      - users
  /users/refresh-token:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New access token
          schema:
            properties:
              data:
                $ref: '#/definitions/model.UserResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Invalid refresh token
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      summary: Refresh access token
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'API token authentication. Format: your-token-here (without "Bearer"
      prefix)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"